tagging CacheNT baseline

+++ MapView/Map/RMMapLayer.m
@@ -1,7 +1,7 @@
 //
 //  RMMapLayer.m
 //
-// Copyright (c) 2008-2009, Route-Me Contributors
+// Copyright (c) 2008, Route-Me Contributors
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -46,7 +46,6 @@
 	return self;
 }
 
-/// \bug why return nil for the "position" and "bounds" actionForKey? Does this do anything besides block Core Animation?
 - (id<CAAction>)actionForKey:(NSString *)key
 {
 	if ([key isEqualToString:@"position"]
@@ -61,7 +60,7 @@
 	self.position = RMTranslateCGPointBy(self.position, delta);
 }
 
-- (void)zoomByFactor: (float) zoomFactor near:(CGPoint) pivot
+- (void)zoomByFactor: (double) zoomFactor near:(CGPoint) pivot
 {
 	self.position = RMScaleCGPointAboutPoint(self.position, zoomFactor, pivot);
 	self.bounds = RMScaleCGRectAboutPoint(self.bounds, zoomFactor, self.anchorPoint);
