+++ MapView/Map/RMPixel.h
@@ -1,7 +1,7 @@
 //
 //  RMPixel.h
 //
-// Copyright (c) 2008-2009, Route-Me Contributors
+// Copyright (c) 2008, Route-Me Contributors
 // All rights reserved.
 //
 // Redistribution and use in source and binary forms, with or without
@@ -32,8 +32,8 @@
 
 #include <CoreGraphics/CGGeometry.h>
 
-CGPoint RMScaleCGPointAboutPoint(CGPoint point, float factor, CGPoint pivot);
-CGRect RMScaleCGRectAboutPoint(CGRect rect, float factor, CGPoint pivot);
+CGPoint RMScaleCGPointAboutPoint(CGPoint point, double factor, CGPoint pivot);
+CGRect RMScaleCGRectAboutPoint(CGRect rect, double factor, CGPoint pivot);
 CGPoint RMTranslateCGPointBy(CGPoint point, CGSize delta);
 CGRect RMTranslateCGRectBy(CGRect rect, CGSize delta);
